requirements.txt
Flask==3.1.1
flask_cors==6.0.0
spotipy==2.25.1
lyricsgenius==3.6.4
transformers==4.41.1
torch>=2.0.0
wordcloud==1.9.3
matplotlib==3.8.4
Pillow==10.3.0
nltk==3.8.1
scikit-learn==1.5.0
numpy
multidict
accelerate
deepl
konlpy
JPype1

firebase-admin
python-dotenv